Create overrides.js
This file is meant to contain json objects that mimic the responses of the Bing Maps REST Locations API - the purpose is to utilize these if certain conditions are met instead of calling the Locations API for the results. This is good practice when the Locations API returns incorrect or unexpected data and a quick data fix is required instead of waiting for the Bing Maps support teams to fix it.