Create overrider.js
This scripts handles the geocoding of the override functionality. It will check the json objects in overrides.js to see if there are any matches to the users' input. If there are, the object is pulled and used instead of calling the Bing Maps API. If there is no match, the Bing Maps API is used.